Move provider helpers into api crates
