+++ crates/google_ai/Cargo.toml
@@ -18,8 +18,11 @@ schemars = ["dep:schemars"]
 anyhow.workspace = true
 futures.workspace = true
 http_client.workspace = true
+language_model_core.workspace = true
+log.workspace = true
 schemars = { workspace = true, optional = true }
 serde.workspace = true
 serde_json.workspace = true
 settings.workspace = true
 strum.workspace = true
+tiktoken-rs.workspace = true
