+++ crates/open_ai/Cargo.toml
@@ -17,9 +17,11 @@ schemars = ["dep:schemars"]
 
 [dependencies]
 anyhow.workspace = true
+collections.workspace = true
 futures.workspace = true
 http_client.workspace = true
 language_model_core.workspace = true
+partial-json-fixer.workspace = true
 rand.workspace = true
 schemars = { workspace = true, optional = true }
 log.workspace = true
@@ -28,3 +30,7 @@ serde_json.workspace = true
 settings.workspace = true
 strum.workspace = true
 thiserror.workspace = true
+tiktoken-rs.workspace = true
+
+[dev-dependencies]
+pretty_assertions.workspace = true
